backend/requirements.txt
# cross fade
pydub
ffmpeg-python

# beat matching
librosa
numpy
scipy

# software
fastapi
uvicorn[standard]
python-multipart
Pillow
requests
aiofiles

# frontend
streamlit



# Experiment notebook 

ipykernel
matplotlib

